refactored test, moved user tests to UserTest.java
